+++ src/endpoints/meta/data/handler.ts
@@ -2,9 +2,8 @@ import {ApiResponseCode, DataPageMetaPayload} from '../../../api-def/api';
 import {processDataMetaPayload} from '../../../utils/payload';
 import {HandlerParams} from '../../lookup';
 import {ApiFailedResponse} from '../../post/base/response/failed';
-import {KeyPointController} from '../../tier/points/controller';
 import {UserController} from '../../userControl/controller';
-import {generateResponse} from '../utils';
+import {dataMetaHandlers} from './handlers';
 import {DataPageMetaResponse} from './response';
 
 
@@ -14,22 +13,11 @@ export const handleDataMeta = async ({
 }: HandlerParams<DataPageMetaPayload>): Promise<DataPageMetaResponse | ApiFailedResponse> => {
   payload = processDataMetaPayload(payload);
 
-  const userData = await UserController.getUserData(mongoClient, payload.uid);
-
-  if (payload.type === 'tierKeyPoint') {
-    const title = await KeyPointController.getDescription(mongoClient, payload.lang, payload.id);
-
-    if (!title) {
-      return new ApiFailedResponse(ApiResponseCode.FAILED_DATA_NOT_EXISTS);
-    }
-
-    return await generateResponse(
-      payload,
-      mongoClient,
-      userData,
-      (options) => new DataPageMetaResponse({...options, params: {title}}),
-    );
+  if (!Object.keys(dataMetaHandlers).includes(payload.type)) {
+    return new ApiFailedResponse(ApiResponseCode.FAILED_UNHANDLED_DATA_TYPE);
   }
 
-  return new ApiFailedResponse(ApiResponseCode.FAILED_UNHANDLED_DATA_TYPE);
+  const user = await UserController.getUserData(mongoClient, payload.uid);
+
+  return dataMetaHandlers[payload.type]({mongoClient, payload, user});
 };
