+++ src/endpoints/meta/data/handlers/index.ts
@@ -1,0 +1,10 @@
+import {DataType} from '../../../../api-def/api';
+import {DataMetaHandler} from '../type';
+import {datamineHandler} from './datamine';
+import {tierKeyPointHandler} from './tierKeyPoint';
+
+
+export const dataMetaHandlers: {[type in DataType]: DataMetaHandler} = {
+  tierKeyPoint: tierKeyPointHandler,
+  datamine: datamineHandler,
+};
